requirements.txt
PyCryptodome
mitmproxy